Move blower_fsm implementations into cpp file.
Main motivation is to allow us to declare constants and functions that
are "private" to the file.

GPC: move-fsm-impls
